FGSA-CB/requirements.txt
labelme==5.2.1
matplotlib==3.1.2
numpy==1.23.3
onnx==1.14.0
onnxsim==0.4.33
opencv_python==4.6.0.66
opencv_python_headless==4.7.0.72
Pillow==9.4.0
Pillow==10.0.0
pytorch_lightning==1.1.8
scipy==1.9.1
segmentation_models_pytorch==0.3.0
thop==0.1.1.post2209072238
torch==1.12.1
torchsummary==1.5.1
torchvision==0.13.1
tqdm==4.64.1
tqdm==4.65.0
